+++ 2024/10/b.cpp
@@ -26,8 +26,8 @@ int main() {
   }
 
   int totalWays = 0;
-  int rows = grid.size();
-  int cols = grid[0].size();
+  int rows = static_cast<int>(grid.size());
+  int cols = static_cast<int>(grid[0].size());
   vector<vector<int>> ways(rows, vector<int>(cols, 0)); // DP table to store path counts
 
   for (auto [x, y] : zeroPositions) {
